+++ PriceAggregator.API/Services/IPriceAggregationStrategy.cs
@@ -2,5 +2,5 @@ namespace PriceAggregator.API.Services;
 
 public interface IPriceAggregationStrategy
 {
-    decimal AggregatePrices(IReadOnlyList<decimal> prices);
+    double AggregatePrices(IReadOnlyList<double> prices);
 }
