+++ PriceAggregator.API/Services/IPriceAggregator.cs
@@ -2,5 +2,5 @@ namespace PriceAggregator.API.Services;
 
 public interface IPriceAggregator
 {
-    Task<decimal> AggregatePrice(IReadOnlyList<IPriceSource> priceSources, string financialInstrument, DateTime time);
+    Task<double> AggregatePrice(IReadOnlyList<IPriceSource> priceSources, string financialInstrument, DateTime time);
 }
